refactor: extract shared task list component
